# PoC: Add Predicate Pushdown to Parquet Reader for Optimized Query Performance