@ethe 2025-03-31
> I think it is possible to implement this feature without modifing the parquet reader and using the currently available APIs

I have tried to implement this in third-party libs, but arrow-rs lacks enough public APIs (for example, users can not construct `Sbbf` outside of `parquet`), also the related APIs is not convenient enough to be used in public at the moment.

> That being said, as you show here it is non trivial to implement row group / page filtering.

That's what I want to point out, this demand is general enough to lots of users, but is not that easy to be realized, if parquet contains a first-party `TableProvider` implementation, it is good to me.