@ethe 2025-03-31
> What is the use case for constructing Sbbf?

I can not find the way to get `Sbbf` instances in the async read path of `parquet` crates, this only works with `SerializedRowGroupReader`, but it is synchronous, so I have to construct it manually from `bytes::Bytes`. 

> If you are using DataFusion already

I do not use datafusion (not yet), if there is a first-party scan method of parquet async reader with prediction/projection/limitation pushdown, that is what I need. I'd like to say `TableProvider` provides similar semantics to the above API, but I'm not sure it is the best choice to be the first-party implementation in `parquet`.

> My biggest concern here is adding more code to maintain as part of this crate that may not be widely used

Tonbo and also Chroma(@HammadB) both run into this issue.