@alamb 2025-04-01
Thanks @HammadB  -- I tried to improve the documentation here too, let me know what you think:
- https://github.com/apache/arrow-rs/pull/7370